Revert to using Keras Sequence for TensorFlow Loader (#79)
* Revert to using Keras Sequence for TensorFlow Loader

Using with the generator data adapter seems to only support a single run through the
dataset. (1 epoch).

Switching back to the keras sequence for now to preserve existing epoch functionality.

* Add assert for loss values from tf model test.

This will catch a case where the second epoch doesn't loop through the dataset a second
time. When using it as an iterator for example.

* Add keras Sequence to Loader class

* Remove tests of peek and restore

* Revert change to LoaderBase.__iter__ to reset iter on call

Co-authored-by: edknv <109497216+edknv@users.noreply.github.com>